refactor(ui5-split-button): migrate to TS
