+++ packages/fiori/src/Interfaces.ts
